requirements.txt
python-telegram-bot==20.7
openai==1.12.0
configparser==6.0.0
PyYAML==6.0.1
python-dotenv==1.0.1
